refactored components in seperate files
