+++ src/components/CategoryPanel/CategoryItem.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+import { Link } from "react-router-dom";
+
+const CategoryItem = ({ title, categoryUrl }) => {
+  return (
+    <li>
+      <Link to={categoryUrl}>{title}</Link>
+    </li>
+  );
+};
+
+export default CategoryItem;
