+++ src/components/Footer/index.js
@@ -1,54 +1,35 @@
 import React from "react";
+import FooterItem from "./FooterItem";
 
 const Footer = () => {
   return (
     <footer className="footer">
       <div className="container">
         <div className="columns is-multiline">
-          <div className="column">
-            <nav className="menu">
-              <h2 className="menu-label">Categories</h2>
-              <ul className="menu-list">
-                <li>Item</li>
-                <li>Item</li>
-                <li>Item</li>
-                <li>Item</li>
-              </ul>
-            </nav>
-          </div>
-          <div className="column">
-            <nav className="menu">
-              <h2 className="menu-label">Categories</h2>
-              <ul className="menu-list">
-                <li>Item</li>
-                <li>Item</li>
-                <li>Item</li>
-                <li>Item</li>
-              </ul>
-            </nav>
-          </div>
-          <div className="column">
-            <nav className="menu">
-              <h2 className="menu-label">Categories</h2>
-              <ul className="menu-list">
-                <li>Item</li>
-                <li>Item</li>
-                <li>Item</li>
-                <li>Item</li>
-              </ul>
-            </nav>
-          </div>
-          <div className="column">
-            <nav className="menu">
-              <h2 className="menu-label">Categories</h2>
-              <ul className="menu-list">
-                <li>Item</li>
-                <li>Item</li>
-                <li>Item</li>
-                <li>Item</li>
-              </ul>
-            </nav>
-          </div>
+          <FooterItem title="Categories">
+            <FooterItem.Link to="/">Link 1</FooterItem.Link>
+            <FooterItem.Link to="/">Link 2</FooterItem.Link>
+            <FooterItem.Link to="/">Link 3</FooterItem.Link>
+            <FooterItem.Link to="/">Link 4</FooterItem.Link>
+          </FooterItem>
+          <FooterItem title="Categories">
+            <FooterItem.Link to="/">Link 1</FooterItem.Link>
+            <FooterItem.Link to="/">Link 2</FooterItem.Link>
+            <FooterItem.Link to="/">Link 3</FooterItem.Link>
+            <FooterItem.Link to="/">Link 4</FooterItem.Link>
+          </FooterItem>
+          <FooterItem title="Categories">
+            <FooterItem.Link to="/">Link 1</FooterItem.Link>
+            <FooterItem.Link to="/">Link 2</FooterItem.Link>
+            <FooterItem.Link to="/">Link 3</FooterItem.Link>
+            <FooterItem.Link to="/">Link 4</FooterItem.Link>
+          </FooterItem>
+          <FooterItem title="Categories">
+            <FooterItem.Link to="/">Link 1</FooterItem.Link>
+            <FooterItem.Link to="/">Link 2</FooterItem.Link>
+            <FooterItem.Link to="/">Link 3</FooterItem.Link>
+            <FooterItem.Link to="/">Link 4</FooterItem.Link>
+          </FooterItem>
         </div>
       </div>
     </footer>
