+++ src/components/Header/index.js
@@ -1,26 +1,7 @@
 import React from "react";
+import NavbarLink from "./NavbarLink";
 import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
-
-const NavbarLink = ({ url, title }) => {
-  return (
-    <li className="navbar-item">
-      <Link to={url} className="navbar-link is-arrowless">
-        {title}
-      </Link>
-    </li>
-  );
-};
-
-NavbarLink.propTypes = {
-  url: PropTypes.string.isRequired,
-  title: PropTypes.string.isRequired,
-};
-
-NavbarLink.defaultProps = {
-  url: "/",
-  title: "",
-};
 
 export default class Header extends React.Component {
   static Link = NavbarLink;
