+++ src/components/SiteArea/LeftSidebar.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const LeftSidebar = ({ children }) => {
+  return <aside className="column is-one-quarter mx-2 my-2">{children}</aside>;
+};
+
+export default LeftSidebar;
